pulled createSrcset into helpers, updated snapshots
