+++ src/app/containers/Image/index.jsx
@@ -7,6 +7,7 @@ import {
   GridItemConstrainedMedium,
   GridItemConstrainedSmall,
 } from '../../lib/styledGrid';
+import createSrcset from './helpers/srcSetHelper';
 
 const DEFAULT_IMAGE_RES = 640;
 
@@ -29,25 +30,6 @@ const getIChefURL = (originCode, locator) => {
 
 const getRawImageSrc = (originCode, locator) =>
   originCode !== 'pips' ? getIChefURL(originCode, locator) : locator;
-
-const createSrcset = (originCode, locator) => {
-  let imgSet = '';
-  const resolutions = [240, 320, 480, 624, 800];
-  if (originCode !== 'pips') {
-    const overridableOriginCode = originCode || 'cpsdevpb';
-
-    const urls = resolutions.map(
-      resolution =>
-        `https://ichef.bbci.co.uk/news/${resolution}/${overridableOriginCode}/${locator} ${resolution}w`,
-    );
-
-    imgSet = urls.join(', ');
-  } else {
-    imgSet = null;
-  }
-
-  return imgSet;
-};
 
 const ImageContainer = ({ blocks }) => {
   if (!blocks) {
